Updating static resources from the 11.8.1 Datadragon release
New hero Gwen was released, so needed to update the static resoucres for
the project. Right now we're just pulling every image from datadragon,
would like to make an automation script to pull only the assests we need
and to replace the old with the new for when their is a new relase.

Also, I'm keeping the old json file on the git so there's access to it
for checking how the json file is structured. Might be easier than going
from the compressed version from DD, to prettyprinted, back to
compressed to save space on heroku slugs.