update_match will insert missing match history for anyone in sum table
Update match is now able to handle summoners that don't have a match
history and have just been added to our summoner table. It should also
handle interuptions better as the games are now inserted 'from the
bottom up'. So if a call to update_match call is interupted, or the
process is stopped early, we don't get gaps in match history.